v2.3.1
小改动

+++ src/views/AllChats.vue
@@ -28,7 +28,7 @@
         <el-table-column prop="did" label="DID" width="80" />
         <el-table-column prop="clientId" label="客服id" width="80" />
         <el-table-column prop="consumerId" label="顾客id" width="80" />
-        <el-table-column prop="content" label="内容" />
+        <el-table-column prop="content" label="内容" width="240" />
         <el-table-column prop="role" label="角色" width="80" />
         <el-table-column prop="sensitiveReason" label="敏感词" width="80" />
         <!-- <el-table-column prop="createTime" label="创建时间" width="160" /> -->
@@ -39,7 +39,7 @@
             <el-button type="primary" size="small" @click="goEdit(scope.$index, scope.row)">
               编辑
             </el-button>
-            <el-button type="danger" size="small" @click="goDelete(scope.$index, scope.row)">
+            <el-button type="danger" size="small" @click="confirmDelete(scope.$index, scope.row)">
               删除
             </el-button>
           </template>
@@ -85,7 +85,9 @@
 import { ref, onMounted } from 'vue'
 import { getChatPage, deleteByCid, updateByCid } from '@/apis/chatApi'
 import type { DialogueItem } from '@/stores/types'
-
+import { ElMessageBox } from 'element-plus'
+import 'element-plus/theme-chalk/el-overlay.css';
+import 'element-plus/theme-chalk/el-message-box.css';
 
 const sendData = ref({ pageSize: 10, currentPage: 1, type: '', searchContent: null })
 let contentData = ref<DialogueItem[]>([])
@@ -205,13 +207,28 @@ const submitEdit = async () => {
 
 }
 
-const goDelete = async (index: any, row: any) => {
-  try {
-    const res = await deleteByCid({ cid: row.cid })
-    ElMessage.success('删除成功！')
-    handleSearch()
-  } catch (error) {
-  }
+const confirmDelete = (index: any, row: any) => {
+  ElMessageBox.confirm(
+    '确认删除这条记录吗？操作不可撤销。',
+    '提示',
+    {
+      confirmButtonText: '确定',
+      cancelButtonText: '取消',
+      type: 'warning',
+    }
+  )
+    .then(async () => {
+      try {
+        const res = await deleteByCid({ cid: row.cid })
+        ElMessage.success('删除成功')
+        handleSearch() // 重新刷新列表
+      } catch (error) {
+        ElMessage.error('删除失败')
+      }
+    })
+    .catch(() => {
+      ElMessage.info('已取消删除')
+    })
 }
 
 const formatDate = (str: string) => {
